refactor: cmsParagraph read actions
