+++ src/actions/cms/paragraphs/read.ts
@@ -5,48 +5,29 @@ import prisma from '@/prisma'
 import { SpecialCmsParagraph } from '@prisma/client'
 import type { ActionReturn } from '@/actions/Types'
 import type { CmsParagraph } from '@prisma/client'
+import { readCmsParagraph, readSpecialCmsParagraph } from '@/server/cms/paragraphs/read'
 
-export async function readCmsParagraph(name: string): Promise<ActionReturn<CmsParagraph>> {
-    try {
-        const paragraph = await prisma.cmsParagraph.findUnique({
-            where: {
-                name
-            }
-        })
-        if (!paragraph) return createActionError('NOT FOUND', 'CmsParagraph not found')
-        return {
-            success: true,
-            data: paragraph
-        }
-    } catch (error) {
-        return createPrismaActionError(error)
-    }
+export async function readCmsParagraphAction(name: string): Promise<ActionReturn<CmsParagraph>> {
+    const paragraphRes = await readCmsParagraph(name)
+    //TODO: Auth on visibility (or permission if special)
+    return paragraphRes
 }
 
 /**
- * This function reads a special paragraph from the database, if it does not exist it will create it
+ * This action reads a special paragraph from the database, if it does not exist it will create it
  * @param special - special paragraph to read
  * @returns - the paragraph
  */
-export async function readSpecialCmsParagraph(special: SpecialCmsParagraph): Promise<ActionReturn<CmsParagraph>> {
+export async function readSpecialCmsParagraphAction(special: SpecialCmsParagraph): Promise<ActionReturn<CmsParagraph>> {
     if (!Object.values(SpecialCmsParagraph).includes(special)) {
         return createActionError('BAD PARAMETERS', `${special} is not special`)
     }
 
-    try {
-        const paragraph = await prisma.cmsParagraph.findUnique({
-            where: {
-                special
-            }
-        })
-        if (paragraph) {
-            return {
-                success: true,
-                data: paragraph
-            }
-        }
-        return createCmsParagraph(special, { special })
-    } catch (error) {
-        return createPrismaActionError(error)
+    const specialRes = await readSpecialCmsParagraph(special)
+    if (specialRes.success) return specialRes
+    if (specialRes.errorCode === 'NOT FOUND') {
+        const created = await createCmsParagraph(special, { special })
+        return created
     }
+    return specialRes
 }
