+++ src/app/components/Cms/CmsParagraph/SpecialCmsParagraph.tsx
@@ -1,5 +1,5 @@
 import CmsParagraph from './CmsParagraph'
-import { readSpecialCmsParagraph } from '@/actions/cms/paragraphs/read'
+import { readSpecialCmsParagraphAction } from '@/actions/cms/paragraphs/read'
 import React from 'react'
 import type { PropTypes as PropTypesCmsParapraph } from './CmsParagraph'
 import type { SpecialCmsParagraph as SpecialCmsParagraphT } from '@prisma/client'
@@ -9,15 +9,15 @@ type PropTypes = Omit<PropTypesCmsParapraph, 'cmsParagraph'> & {
 }
 
 /**
- * WARNING: This component should only be rendered server side, on client fetch the paragraph using readSpecialCmsParagraph
+ * WARNING: This component should only be rendered server side, on client fetch the paragraph using readSpecialCmsParagraphAction
  * then render it using CmsParagraph.
- * A component that reads a special type of paragraph using readSpecialCmsParagraph
+ * A component that reads a special type of paragraph using readSpecialCmsParagraphAction
  * then renders it using standard CmsParagraph
  * @param special - the special type of the paragraph to read
  * @returns
  */
 export default async function SpecialCmsParagraph({ special, ...props }: PropTypes) {
-    const res = await readSpecialCmsParagraph(special)
+    const res = await readSpecialCmsParagraphAction(special)
     if (!res.success) return <i>Error: {res.error && res.error[0].message}</i>
 
     return (
